have node modules do asbsolute vs relative path checks, refactored logic to resource interface
