+++ packages/cli/src/plugins/resource/plugin-node-modules.js
@@ -178,8 +178,11 @@ class NodeModulesResource extends ResourceInterface {
   async resolve(url) {
     return new Promise((resolve, reject) => {
       try {
-        const relativeUrl = url.replace(this.compilation.context.userWorkspace, '');
-        const nodeModulesUrl = path.join(process.cwd(), relativeUrl);
+        const { projectDirectory } = this.compilation.context;
+        const isAbsoluteNodeModulesFile = fs.existsSync(path.join(projectDirectory, url));
+        const nodeModulesUrl = isAbsoluteNodeModulesFile
+          ? path.join(projectDirectory, url)
+          : path.join(projectDirectory, this.getReducedUrl(projectDirectory, url));
 
         resolve(nodeModulesUrl);
       } catch (e) {
