+++ packages/cli/test/cases/build.default.workspace-templates-relative-paths/src/styles/page.css
@@ -1,3 +1,5 @@
+@import url('../../node_modules/prismjs/themes/prism-tomorrow.css');
+
 ul {
   background: purple;
 }
